+++ nodes/property-read.js
@@ -15,14 +15,20 @@ module.exports = function(RED) {
             let assetId = config.assetId || msg.asset_id;
             let property = config.property || msg.property;
 
-            if (typeof server.readProperty === "function") {
-                server.readProperty(asset, assetId, property, function(res) {
-                  msg.payload = res;
-                  send(msg);
-                })
-                .catch(e => node.error(e));
-            } else
-                node.error("Check Thinger Server Configuration");
+            const method = 'GET';
+            const apiVersion = (asset == "devices" ? "v3" : "v1");
+            const url = `${server.config.ssl ? "https://" : "http://"}${server.config.host}/${apiVersion}/users/${server.config.username}/${asset}/${assetId}/properties/${property}`;
+
+            if (typeof server.request === "function") {
+              server.request(node, url, method)
+              .then(res => {
+                  msg.payload = res.payload;
+                  node.send(msg);
+              })
+              .catch(e => node.error(e));
+            }
+            else
+              node.error("Check Thinger Server Configuration");
         });
     }