+++ nodes/property-write.js
@@ -10,23 +10,47 @@ module.exports = function(RED) {
         var server = RED.nodes.getNode(config.server);
 
         // call property read on input
-        node.on("input",function(msg, send) {
+        node.on("input",async function(msg, send) {
             let asset = (config.asset || msg.asset)+"s";
             let assetId = config.assetId || msg.asset_id;
             let property = config.property || msg.property;
-            var value = config.value || msg.payload || msg.value;
-            if (typeof(value) === 'object') {
-                value = JSON.stringify(value);
-            }
+            var data = config.value || msg.payload || msg.value;
 
-            if (typeof server.writeProperty === "function") {
-                server.writeProperty(asset, assetId, property, value, function(res) {
-                  msg.payload = res;
-                  send(msg);
-                })
-                .catch(e => node.error(e));
-            } else
-                node.error("Check Thinger Server Configuration");
+            const apiVersion = (asset == "devices" ? "v3" : "v1");
+            const url = `${server.config.ssl ? "https://" : "http://"}${server.config.host}/${apiVersion}/users/${server.config.username}/${asset}/${assetId}/properties`;
+
+            if (typeof server.request === "function") {
+                // Check if property exists
+                let exists = false;
+                let res;
+                res = await server.request(node,`${url}`, 'GET');
+                for (let i in res) {
+                    if (res[i].property === property) {
+                        exists = true;
+                        break;
+                    }
+                }
+
+                data = {property: property, value: JSON.parse(data)};
+
+                // Update if exist or create it
+                let method;
+                if ( exists )
+                    method = 'PUT';
+                else
+                    method = 'POST';
+                try {
+                    res = await server.request(node, url, method, data);
+                } catch (err) {
+                    node.error(err);
+                    return;
+                }
+
+                msg.payload = res.payload;
+                send(msg);
+            }
+            else
+              node.error("Check Thinger Server Configuration");
         });
     }